Split up giant formSchemas file, into multi files
Splitting them out to multiple files could arguably make it
easier to delete/mod certain parts of the app
