+++ schemas/confirmation.schema.js
@@ -1,0 +1,12 @@
+const reviewSchema = {
+  review: {
+    isIn: {
+      errorMessage: 'errors.review',
+      options: [['review']],
+    },
+  },
+}
+
+module.exports = {
+  reviewSchema
+}
