+++ schemas/financial.schema.js
@@ -1,0 +1,9 @@
+const { yesNoSchema } = require('./utils.schema')
+
+const incomeSchema = {
+  confirmIncome: yesNoSchema(),
+}
+
+module.exports = {
+  incomeSchema,
+}
